Added project files for motion_detect branch
New set of files continually captures & deletes photos until motion is detected. At which point a video is recorded for the next minute. The photos that triggered the video and the video itself are saved on  the local machine, and another program uses sockets to transmit them to a base station for processing.